requirements.txt
pandas
Pillow
tqdm
numpy
scikit-learn
tensorboardX
tensorboard
scipy
tensorflow
torchvision
huggingface_hub
opencv-python-headless
scikit-image
git+https://github.com/mttgdd/bc_radar_vgg19.git